2.0.6a -
The Blood Magic Life Shard was changed from infusing a weak blood shard with 35 000 LP on a Tier 3 Blood Altar to an Extra Life from the  BetterQuestingMod infused with 50 000 LP on a Tier 5 Blood Altar.

The quest for the Life Shard was moved from the tier 3 Blood Altar quests to the tier 5 Blood Altar quests.